+++ pcs/config.py
@@ -20,6 +20,7 @@ import tempfile
 import time
 import platform
 import shutil
+import difflib
 
 try:
     import clufter.facts
@@ -33,7 +34,6 @@ except ImportError:
 from pcs import (
     cluster,
     constraint,
-    prop,
     quorum,
     resource,
     settings,
@@ -43,12 +43,16 @@ from pcs import (
     utils,
     alert,
 )
+from pcs.cli.common import middleware
+from pcs.cli.common.console_report import indent
+from pcs.cli.constraint import command as constraint_command
+from pcs.cli.constraint_colocation import (
+    console_report as colocation_console_report,
+)
+from pcs.cli.constraint_order import console_report as order_console_report
+from pcs.cli.constraint_ticket import console_report as ticket_console_report
 from pcs.lib.errors import LibraryError
 from pcs.lib.commands import quorum as lib_quorum
-import pcs.cli.constraint_colocation.command as colocation_command
-import pcs.cli.constraint_order.command as order_command
-import pcs.cli.constraint_ticket.command as ticket_command
-from pcs.cli.common.console_report import indent
 
 
 def config_cmd(argv):
@@ -72,6 +76,8 @@ def config_cmd(argv):
             config_checkpoint_view(argv[1:])
         elif argv[0] == "restore":
             config_checkpoint_restore(argv[1:])
+        elif argv[0] == "diff":
+            config_checkpoint_diff(argv[1:])
         else:
             usage.config(["checkpoint"])
             sys.exit(1)
@@ -93,10 +99,11 @@ def config_cmd(argv):
         sys.exit(1)
 
 def config_show(argv):
+    lib = utils.get_library_wrapper()
     print("Cluster Name: %s" % utils.getClusterName())
     status.nodes_status(["config"])
     print()
-    config_show_cib()
+    print("\n".join(_config_show_cib_lines(lib)))
     if (
         utils.hasCorosyncConf()
         and
@@ -122,42 +129,95 @@ def config_show(argv):
         except LibraryError as e:
             utils.process_library_reports(e.args)
 
-def config_show_cib():
-    lib = utils.get_library_wrapper()
-    modifiers = utils.get_modifiers()
-
-    print("Resources:")
+def _config_show_cib_lines(lib):
+    # update of pcs_options will change output of constraint show
     utils.pcs_options["--all"] = 1
     utils.pcs_options["--full"] = 1
-    resource.resource_show([])
+    # get latest modifiers object after updating pcs_options
+    modifiers = utils.get_modifiers()
 
-    print()
-    print("Stonith Devices:")
-    resource.resource_show([], True)
-    print("Fencing Levels:")
-    levels = stonith.stonith_level_config_to_str(
+    cib_xml = utils.get_cib()
+    cib_etree = utils.get_cib_etree(cib_xml=cib_xml)
+    cib_dom = utils.get_cib_dom(cib_xml=cib_xml)
+
+    resource_lines = []
+    stonith_lines = []
+    for resource_el in cib_etree.find(".//resources"):
+        is_stonith = (
+            "class" in resource_el.attrib
+            and
+            resource_el.attrib["class"] == "stonith"
+        )
+        resource_el_lines = resource.resource_node_lines(resource_el)
+        if is_stonith:
+            stonith_lines += resource_el_lines
+        else:
+            resource_lines += resource_el_lines
+
+    all_lines = []
+
+    all_lines.append("Resources:")
+    all_lines.extend(indent(resource_lines, indent_step=1))
+    all_lines.append("")
+    all_lines.append("Stonith Devices:")
+    all_lines.extend(indent(stonith_lines, indent_step=1))
+    all_lines.append("Fencing Levels:")
+    levels_lines = stonith.stonith_level_config_to_str(
         lib.fencing_topology.get_config()
+             )
+    if levels_lines:
+        all_lines.extend(indent(levels_lines, indent_step=2))
+
+    all_lines.append("")
+    constraints_element = cib_dom.getElementsByTagName('constraints')[0]
+    all_lines.extend(
+        constraint.location_lines(constraints_element, showDetail=True)
     )
-    if levels:
-        print("\n".join(indent(levels, 2)))
+    all_lines.extend(constraint_command.show(
+        "Ordering Constraints:",
+        lib.constraint_order.show,
+        order_console_report.constraint_plain,
+        modifiers
+    ))
+    all_lines.extend(constraint_command.show(
+         "Colocation Constraints:",
+        lib.constraint_colocation.show,
+        colocation_console_report.constraint_plain,
+        modifiers
+    ))
+    all_lines.extend(constraint_command.show(
+        "Ticket Constraints:",
+        lib.constraint_ticket.show,
+        ticket_console_report.constraint_plain,
+        modifiers
+    ))
 
-    print()
-    constraint.location_show([])
-    order_command.show(lib, [], modifiers)
-    colocation_command.show(lib, [], modifiers)
-    ticket_command.show(lib, [], modifiers)
+    all_lines.append("")
+    all_lines.extend(alert.alert_config_lines(lib))
 
-    print()
-    alert.print_alert_config(lib, [], modifiers)
+    all_lines.append("")
+    all_lines.append("Resources Defaults:")
+    all_lines.extend(indent(
+        resource.show_defaults(cib_dom, "rsc_defaults"),
+        indent_step=1
+    ))
+    all_lines.append("Operations Defaults:")
+    all_lines.extend(indent(
+        resource.show_defaults(cib_dom, "op_defaults"),
+        indent_step=1
+    ))
 
-    print()
-    del utils.pcs_options["--all"]
-    print("Resources Defaults:")
-    resource.show_defaults("rsc_defaults", indent=" ")
-    print("Operations Defaults:")
-    resource.show_defaults("op_defaults", indent=" ")
-    print()
-    prop.list_property([])
+    all_lines.append("")
+    all_lines.append("Cluster Properties:")
+    properties = utils.get_set_properties()
+    all_lines.extend(indent(
+        [
+            "{0}: {1}".format(prop, val)
+            for prop, val in sorted(properties.items())
+        ],
+        indent_step=1
+    ))
+    return all_lines
 
 def config_backup(argv):
     if len(argv) > 1:
@@ -588,20 +648,87 @@ def config_checkpoint_list():
             % (cib_info[1], datetime.datetime.fromtimestamp(round(cib_info[0])))
         )
 
+def _checkpoint_to_lines(lib, checkpoint_number):
+    # backup current settings
+    orig_usefile = utils.usefile
+    orig_filename = utils.filename
+    orig_middleware = lib.middleware_factory
+    # configure old code to read the CIB from a file
+    utils.usefile = True
+    utils.filename = os.path.join(
+        settings.cib_dir,
+        "cib-%s.raw" % checkpoint_number
+    )
+    # configure new code to read the CIB from a file
+    lib.middleware_factory = orig_middleware._replace(
+        cib=middleware.cib(utils.filename, utils.touch_cib_file)
+    )
+    # export the CIB to text
+    result = False, []
+    if os.path.isfile(utils.filename):
+        result = True, _config_show_cib_lines(lib)
+    # restore original settings
+    utils.usefile = orig_usefile
+    utils.filename = orig_filename
+    lib.middleware_factory = orig_middleware
+    return result
+
 def config_checkpoint_view(argv):
     if len(argv) != 1:
-        usage.config(["checkpoint", "view"])
+        usage.config(["checkpoint view"])
         sys.exit(1)
 
-    utils.usefile = True
-    utils.filename = os.path.join(settings.cib_dir, "cib-%s.raw" % argv[0])
-    if not os.path.isfile(utils.filename):
+    lib = utils.get_library_wrapper()
+    loaded, lines = _checkpoint_to_lines(lib, argv[0])
+    if not loaded:
         utils.err("unable to read the checkpoint")
-    config_show_cib()
+    print("\n".join(lines))
+
+def config_checkpoint_diff(argv):
+    if len(argv) != 2:
+        usage.config(["checkpoint diff"])
+        sys.exit(1)
+
+    if argv[0] == argv[1]:
+        utils.err("cannot diff a checkpoint against itself")
+
+    lib = utils.get_library_wrapper()
+    errors = []
+    checkpoints_lines = []
+    for checkpoint in argv:
+        if checkpoint == "live":
+            lines = _config_show_cib_lines(lib)
+            if not lines:
+                errors.append("unable to read live configuration")
+            else:
+                checkpoints_lines.append(lines)
+        else:
+            loaded, lines = _checkpoint_to_lines(lib, checkpoint)
+            if not loaded:
+                errors.append(
+                    "unable to read checkpoint '{0}'".format(checkpoint)
+                )
+            else:
+                checkpoints_lines.append(lines)
+
+    if errors:
+        utils.err("\n".join(errors))
+
+    print("Differences between {0} (-) and {1} (+):".format(*[
+        "live configuration" if label == "live"
+            else "checkpoint {0}".format(label)
+        for label in argv
+    ]))
+    print("\n".join([
+        line.rstrip() for line in difflib.Differ().compare(
+            checkpoints_lines[0],
+            checkpoints_lines[1]
+        )]
+    ))
 
 def config_checkpoint_restore(argv):
     if len(argv) != 1:
-        usage.config(["checkpoint", "restore"])
+        usage.config(["checkpoint restore"])
         sys.exit(1)
 
     cib_path = os.path.join(settings.cib_dir, "cib-%s.raw" % argv[0])
@@ -843,7 +970,7 @@ def config_export_pcs_commands(argv, verbose=False):
             invalid_args = True
     # check options
     if invalid_args:
-        usage.config(["export", "pcs-commands"])
+        usage.config(["export pcs-commands"])
         sys.exit(1)
     # complete optional options
     if dist is None: