+++ pcs/constraint.py
@@ -528,7 +528,7 @@ def order_find_duplicates(dom, constraint_el):
 
 # Show the currently configured location constraints by node or resource
 def location_show(argv):
-    if (len(argv) != 0 and argv[0] == "nodes"):
+    if argv and argv[0] == "nodes":
         byNode = True
         showDetail = False
     elif "--full" in utils.pcs_options:
@@ -554,14 +554,27 @@ def location_show(argv):
         valid_noderes = []
 
     (dummy_dom,constraintsElement) = getCurrentConstraints()
+    print("\n".join(location_lines(
+        constraintsElement,
+        showDetail=showDetail,
+        byNode=byNode,
+        valid_noderes=valid_noderes
+    )))
+
+def location_lines(
+    constraintsElement, showDetail=False, byNode=False, valid_noderes=None
+):
+    all_lines = []
     nodehashon = {}
     nodehashoff = {}
     rschashon = {}
     rschashoff = {}
     ruleshash = defaultdict(list)
-    all_loc_constraints = constraintsElement.getElementsByTagName('rsc_location')
+    all_loc_constraints = constraintsElement.getElementsByTagName(
+        'rsc_location'
+    )
 
-    print("Location Constraints:")
+    all_lines.append("Location Constraints:")
     for rsc_loc in all_loc_constraints:
         if rsc_loc.hasAttribute("rsc-pattern"):
             lc_rsc_type = RESOURCE_TYPE_REGEXP
@@ -596,7 +609,7 @@ def location_show(argv):
         else:
             positive = False
 
-        if positive == True:
+        if positive:
             nodeshash = nodehashon
             rschash = rschashon
         else:
@@ -634,12 +647,12 @@ def location_show(argv):
         )
     )
 
-    if byNode == True:
+    if byNode:
         for node in nodelist:
-            if len(valid_noderes) != 0:
+            if valid_noderes:
                 if node not in valid_noderes:
                     continue
-            print("  Node: " + node)
+            all_lines.append("  Node: " + node)
 
             nodehash_label = (
                 (nodehashon, "    Allowed to run:"),
@@ -647,7 +660,7 @@ def location_show(argv):
             )
             for nodehash, label in nodehash_label:
                 if node in nodehash:
-                    print(label)
+                    all_lines.append(label)
                     for options in nodehash[node]:
                         line_parts = [(
                             "      " + options["rsc_label"]
@@ -664,14 +677,14 @@ def location_show(argv):
                                 )
                             )
                         line_parts.append("Score: " + options["score"])
-                        print(" ".join(line_parts))
-        show_location_rules(ruleshash, showDetail)
+                        all_lines.append(" ".join(line_parts))
+        all_lines += _show_location_rules(ruleshash, showDetail)
     else:
         for rsc in rsclist:
-            if len(valid_noderes) != 0:
+            if valid_noderes:
                 if rsc[0:2] not in valid_noderes:
                     continue
-            print("  {0}".format(rsc[2]))
+            all_lines.append("  {0}".format(rsc[2]))
             rschash_label = (
                 (rschashon, "    Enabled on:"),
                 (rschashoff, "    Disabled on:"),
@@ -698,12 +711,14 @@ def location_show(argv):
                             )
                         if showDetail:
                             line_parts.append("(id:{0})".format(options["id"]))
-                        print(" ".join(line_parts))
-            miniruleshash={}
+                        all_lines.append(" ".join(line_parts))
+            miniruleshash = {}
             miniruleshash[rsc] = ruleshash[rsc]
-            show_location_rules(miniruleshash, showDetail, True)
+            all_lines += _show_location_rules(miniruleshash, showDetail, True)
+    return all_lines
 
-def show_location_rules(ruleshash, showDetail, noheader=False):
+def _show_location_rules(ruleshash, showDetail, noheader=False):
+    all_lines = []
     constraint_options = {}
     for rsc in sorted(
         ruleshash.keys(),
@@ -717,7 +732,7 @@ def show_location_rules(ruleshash, showDetail, noheader=False):
     ):
         constrainthash = defaultdict(list)
         if not noheader:
-            print("  {0}".format(rsc[2]))
+            all_lines.append("  {0}".format(rsc[2]))
         for rule in ruleshash[rsc]:
             constraint_id = rule.parentNode.getAttribute("id")
             constrainthash[constraint_id].append(rule)
@@ -731,11 +746,14 @@ def show_location_rules(ruleshash, showDetail, noheader=False):
             else:
                 constraint_option_info = ""
 
-            print("    Constraint: " + constraint_id + constraint_option_info)
+            all_lines.append(
+                "    Constraint: " + constraint_id + constraint_option_info
+            )
             for rule in constrainthash[constraint_id]:
-                print(rule_utils.ExportDetailed().get_string(
+                all_lines.append(rule_utils.ExportDetailed().get_string(
                     rule, showDetail, "      "
                 ))
+    return all_lines
 
 def location_prefer(argv):
     rsc = argv.pop(0)